pkg_getopt_long: use uint64_t for options bitfield

+++ getopt_long.h
@@ -54,7 +54,7 @@ struct pkg_option {
 	 */
 	int has_arg;
 	/* if not NULL, set *flag to val when option found */
-	int *flag;
+	uint64_t *flag;
 	/* if flag not NULL, value to set *flag to; else return value */
 	int val;
 };
